% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data.R
\name{as_data}
\alias{as_data}
\title{transform Raster* to data frame ready for ggplot2}
\usage{
as_data(r, na.rm = FALSE, melt = TRUE)
}
\arguments{
\item{r}{\code{Raster*} object transform}

\item{na.rm}{\code{boolean} should NA pixels be removed}

\item{melt}{\code{boolean} should the raster be melted (default is TRUE)}
}
\value{
a data frame containing \code{x}, \code{y}, the coordinates;  \code{variable}, the layer names; \code{value} the pixel value.
}
\description{
transform Raster* to data frame ready for ggplot2
}
\details{
\code{melt} is used to facilitate ploting of raster stacks
}

