% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{scale_fill_continuous}
\alias{scale_fill_continuous}
\title{Smooth gradient between two colours}
\usage{
scale_fill_continuous(low = "#FFFFB3", high = "#FB8072", ...)
}
\arguments{
\item{low}{Colour for low end of the gradient}

\item{high}{Colour for high end of the gradient}

\item{...}{Other arguments passed on to \code{\link{scale_fill_continuous}}}
}
\description{
Overrides the default continuous gradient as a yellow to red
}

