% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{scale_fill_diverge}
\alias{scale_fill_diverge}
\title{Smooth gradient for diverging data}
\usage{
scale_fill_diverge(low = "#D53E4F", high = "#3288BD", center = "#FFFFBF",
  ...)
}
\arguments{
\item{low}{Colour for low end of the gradient}

\item{high}{Colour for high end of the gradient}

\item{center}{Colour for center of the gradient (default very close to white)}

\item{...}{Other arguments passed on to \code{\link{scale_fill_gradientn}}}
}
\description{
Smooth gradient for diverging data
}

