\name{theme_fullframe}
\alias{theme_fullframe}
\title{A minimalist ggplot2 theme for ploting rasters}
\usage{
  theme_fullframe(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{\code{numeric} base font size}

  \item{base_family}{\code{character} base font family}
}
\value{
  An object of class \code{\link{theme}}.
}
\description{
  Adapted from
  \href{https://github.com/hadley/ggplot2/wiki/Themes}{ggplot2
  wiki}
}

