% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualraster.R
\docType{package}
\name{visualraster}
\alias{visualraster}
\alias{visualraster-package}
\title{Visual Raster Cheatsheet}
\description{
To run The Visual Raster CheatSheet, run \code{demo("vr-cheatsheet")}. You can also have a look at the html vignette by running \code{vignette("visual-raster-cheatsheet")}. Note that the vignette is in html format, so it might not display correctly. Use \code{browseVignettes(package = "visualraster")} in that case and open the html file in a browser.
}

