% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_raster.R
\name{vr_geom_line}
\alias{vr_geom_line}
\title{Custom visual raster line plot}
\usage{
vr_geom_line(l, size = 2, color = "#11111199", ...)
}
\arguments{
\item{l}{\code{data.frame} containing a \code{x} and \code{y} column}

\item{size}{\code{integer} point size}

\item{color}{\code{color} line colors}

\item{...}{other arguments passed to \code{\link{geom_line}}}
}
\description{
Custom visual raster line plot
}

