% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_raster.R
\name{vr_geom_polygon}
\alias{vr_geom_polygon}
\title{Custom visual raster polygon plot}
\usage{
vr_geom_polygon(p, size = 1, fill = "#22222280", color = "#11111199", ...)
}
\arguments{
\item{p}{\code{data.frame} containing at least \code{x} and \code{y} column 
(see \code{\link{fortify}} for a transforming spatial polygons)}

\item{size}{\code{integer} line size}

\item{fill}{\code{color} fill colors}

\item{color}{\code{color} border colors}

\item{...}{other arguments passed to \code{\link{geom_polygon}}}
}
\description{
Custom visual raster polygon plot
}

