\name{vr_geom_raster}
\alias{vr_geom_raster}
\title{Custom plot for Visual Raster}
\usage{
  vr_geom_raster(r, color = I("#222222"), size = I(0.2),
    ...)
}
\arguments{
  \item{r}{\code{Raster*}}

  \item{color}{\code{color} line colors}

  \item{size}{\code{integer} line size}

  \item{...}{other arguments passed to
  \code{\link{geom_tile}}}
}
\description{
  Uses discrete values (maximum 11)
}

