% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_raster.R
\name{vr_geom_raster_seq}
\alias{vr_geom_raster_seq}
\title{Custom visual raster plot for sequential data}
\usage{
vr_geom_raster_seq(r, color = I("#222222"), size = I(0.2), ...)
}
\arguments{
\item{r}{\code{Raster*}}

\item{color}{\code{color} line colors}

\item{size}{\code{integer} line size}

\item{...}{other arguments passed to \code{\link{geom_tile}}}
}
\description{
Custom visual raster plot for sequential data
}
\details{
if \code{ncell(r) > 5e2} will use geom_raster by default for
performance
}

