% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_fullframe}
\alias{theme_fullframe}
\title{A minimalist ggplot2 theme for ploting rasters
 
Adapted from \href{https://github.com/hadley/ggplot2/wiki/Themes}{ggplot2 wiki}}
\usage{
theme_fullframe(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{\code{numeric} base font size}

\item{base_family}{\code{character} base font family}
}
\value{
An object of class \code{\link{theme}}.
}
\description{
A minimalist ggplot2 theme for ploting rasters
 
Adapted from \href{https://github.com/hadley/ggplot2/wiki/Themes}{ggplot2 wiki}
}

